+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -1,6 +1,6 @@
 package nextstep.mvc;
 
-import nextstep.mvc.tobe.HandlerAdapter;
+import nextstep.mvc.tobe.adapter.HandlerAdapter;
 import nextstep.mvc.tobe.ModelAndView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
