+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/adapter/HandlerExecutionAdapter.java
@@ -1,4 +1,7 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.adapter;
+
+import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.handler.HandlerExecution;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
